# gc_registry/static/openapi.yaml
components:
  schemas:
...
    AccountRead:
      properties:
        id:
          format: uuid
          title: Account Id
          type: string
        account_name:
          title: Account Name
          type: string
        organisation:
          $ref: '#/components/schemas/Organisation'
        id:
          format: uuid
          title: Organisation Id
          type: string
        users:
          items:
            $ref: '#/components/schemas/User'
          title: Users
          type: array
      required:
      - account_name
      - id
      - id
      - organisation
      - users
      title: AccountRead
      type: object
    AccountUpdate:
      properties:
        id:
          format: uuid
          title: Account Id
          type: string
        account_name:
          title: Account Name
          type: string
        organisation:
          $ref: '#/components/schemas/Organisation'
        id:
          format: uuid
          title: Organisation Id
          type: string
        id:
          format: uuid
          title: User Id
          type: string
      required:
      - id
      title: AccountUpdate
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    Organisation:
      properties:
        address:
          title: Address
          type: string
        business_id:
          title: Business Id
          type: integer
        name:
          title: Name
          type: string
        id:
          format: uuid
          title: Organisation Id
          type: string
        email:
          title: Primary Contact
          type: string
        website:
          title: Website
          type: string
      required:
      - name
      - business_id
      - email
      title: Organisation
      type: object
...
    User:
      properties:
        name:
          title: Name
          type: string
        id:
          format: uuid
          title: Organisation Id
          type: string
        email:
          title: Primary Contact
          type: string
        role:
          description: "The roles of the User with the registry. A single User can\
            \ be assigned multiple roles\n                       by the Registry Administrator\
            \ (which is itself a User for the purposes of managing allowable\n   \
            \                    actions), including: 'GC Issuer', 'Production Registrar',\
            \ 'Measurement Body', and 'Account Holder'.\n                       The\
            \ roles are used to determine the actions that the User is allowed to\
            \ perform within the registry,\n                       according to the\
            \ EnergyTag Standard."
          items:
            type: string
          title: Role
          type: array
        id:
          format: uuid
          title: User Id
          type: string
      required:
      - name
      - email
      - id
      - role
      title: User
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: token
      type: oauth2
info:
  contact:
    email: connor@futureenergy.associates
    name: Please direct feedback to
  description: "\nThis document outlines the EnergyTag API Specification to accompany\
    \ Version 2 of \nthe EnergyTag Standard.\nThe functionality outlined within each\
    \ API interaction represents the recommended \nminimum required detail necessary\
    \ to implement a consistent Granular Certificate \n(GC) registry system. \nWithin\
    \ the Certificates section, the definition of GCs as specified in the Create \n\
    Certificate POST request, alongside the filtering parameters required to instigate\
    \ \nquery, transfer, and cancellation requests, will form core components of the\
    \ standard \nand impact the functionality of the registry the greatest. Subsequent\
    \ sections \ndescribing Organisation, User, Account, and Device management are\
    \ **recommendations \nonly** and do not form part of the standard, although terminology\
    \ used therein is \nreferenced in the certificate definitions. They are included\
    \ in the database class \ndiagram for the sake of completeness in representing\
    \ the relationships between the \ncertificate and action definitions; the fields\
    \ implemented for these tables are \nleft to the discretion of the registry operator.\
    \ \n<br>\n\nWhere Universally Unique Identifiers (UUIDs) have been used, these\
    \ can be replaced with \nany consistent and unique identification mechanism preferred\
...
openapi: 3.1.0
paths:
...
  /accounts/account/{id}:
    delete:
      operationId: delete_account_accounts_account__id__delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete Account
      tags:
      - Accounts
    get:
      operationId: read_account_accounts_account__id__get
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Read Account
      tags:
      - Accounts
    patch:
      operationId: update_account_accounts_account__id__patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Account
      tags:
      - Accounts
...